Long ago the tag CMD_FAST_EXIT was added to certain bkd commands to indicate that then bkd connection should be terminated after this command.  This was added to support http connections where an EOF on the stream would not occur and we needed to know when a command has completed.  This lack of EOF only happened when talking via a web proxy.
The CMD_FAST_EXIT code is making doing multiple operations in one bkd
connection difficult and so needs to be remoted.

We later had to support proper http headers to work with proxies and
had a context-len header that bkd.c tracks to simulate a EOF when that
number of bytes have been read.  This mostly works but falls down when
the bkd commands themselves consume data from stdin that don't get
counted against the context-len in bkd.c.

Without HTTP only case that I found where we are unintentionally
depending on FAST_EXIT is 'bk pull -n'.  In that case a slight tweak
of the protocol fixes the problem.  This done not cause a
compatibility issue:

   new bk -> old bkd    # the FAST_EXIT will close connection.
   old bk -> new bkd    # the extra @END@ will prevent the deadlock


With HTTP we solve the issue of not seeing a EOF when talking via a
web proxy by having the client side code inject a 'quit\n' and the end
of the stream of commands sent to the bkd.  The command already makes
existing bkd's exit and so it a compatible solution.

Next we had an issue with a BAM pull using HTTP via a web proxy.  The
code used stdio in pull_part2 and yet used direct IO in bkd.c and in
pull_part3.  This worked before because we had a network round trip
between part2 and part3 and so the stdio buffer happened to be empty,
but with this cset using http the stdio buffer contained the 'quit\n'
to close the connection and so the bkd hung.

Now when we start using stdio in part2 a global flag is set and the
bkd.c code will drain the stdio buffer before doing direct IO again.
Also a series of asserts protect all the spaces where we cannot have
buffered IO.

bk: 4ad38bd9gepmq-M2Z6wwzVouum7Jng