Add the ability to run regressions in parallel. Either use 'cd src/t; make -j8' or './doit -j8'
Not used by default cranks.
(We will enable this in the future.)

In order for this cset to work we fixed lots of races and
issues in the regressions and made sure all of bk respects TMPDIR
and won't conflict for files in /tmp.

Part of this was adding 'bk bkd -aPORT' to allocate any free TCP
port and write the one used to the PORT file.  This fixed
serveral random "can't start bkd" failures.

Also this adds a 'bk time' command to time things.

bk: 4ae9b61ezhWwDmUmXEhE6A2RqowHAQ