Second part of the nested locks work.
All lock handling was moved to cmdlog_{start,end} as well as well as
the sending of the server environment if we are in a bkd. This means
all the cmd_*() functions don't need to worry about locking (neither
nested nor repo locking) or handling the environment.

Implement nested read locks and add a 'staling' mechanism that
automatically expires stale locks. A lock is considered stale if the
process that created it is not found (normal case) or if it is an HTTP
lock that hasn't been used for 1/2 hour or was created more than 2
hours ago (hardcoded parameters, will be configurable someday).

Still missing is a plugging in the nested locks into the 'bk lock'
command so that users can see what is going on.

bk: 4af47a3db-R6FAwUQjHauT2hg9YxUA