If we pull/push in a component, turn it into a product pull/push.
This cset introduces BKD_REPOTYPE, which will be sent over the wire
and be set correctly both in the local side and in the remote side.
Most of the protocol decisions that are being taken now by asking
"if (proj_isProduct(0))" should be changed to "if (BKD_REPOTYPE == prod)".
E.g. the decision on whether to unlock remote by sending a "nested unlock"
bkd command.

However, this work was somewhat unrelated to pull/push working on the product
and will thus have its own cset.

bk: 4b1ed47f_lg6YhLB2pvban4kcrQVrw