ADD skipkey handling back in.  The original fastpatch cset dropped skipkey handling in order to get something working. I understand why now -- this adds a bit of ugly to the code.
Skipkeys are created by the lod stripping code (fixlod.c).
The idea is if a repo had lods, strip them out and remember
what was stripped out and in future pulls, ignore them.
The main point here is to just ignore them as though they weren't
there.

To do this in the old code was easy: just ignore.  In the new code,
the delta structure is separated from the interleaved delta, so we
need to remember the ignored delta, and then ignore it in the interleaved
fastpatch handling in slib.c:doFast().  This messes with the whole
pipeline: if ignored, set p->initMmap to 0, in cset_insert, if no
p->initMmap (called iF there), then set serial = 0.  In cset_write,
if serial == 0, add a INVALID to the patchmap addLine of *d.  In
doFast(), if INVALID, ignore any commands and data associated with
it.  Whew.
Thanks to wscott for backporting INVALID and cleanups, especially to
the cweave error path.

bk: 4b323691QqDvTM6oBy8ujnl7rQm7tw