Handle cases where the product's id_cache wasn't always being updated correctly with component ChangeSet files.
Error checks were added to nested_init() to flag cases where
the id_cache was incorrect and then a number of problem
cases were fixed.

The NESTED_FIXIDCACHE flag was added and used in check.c so
that check will repair the id_cache for components like it does
for individual files.

This cset was a joint effort by Rick, Oscar and Wayne

bk: 4b4de82dVWRQZQlgtrjP3bwh6ySkBA