Change how clone behaves with remapped repository. Now clone with preseve the remapping of the source repository instead of always converting to remapped.
A new --sccs-compat & --hide-sccs-dirs options to clone
overrides this default and sets if the SCCS directory
will be visible or not.

A 'bk setup' defaults to creating a remapped directory
but this can be overridden with --sccs-compat.

Fix a bug only seen on ia64 where the locking code would
assert because it was using a stale project*.

bk: 4b632a99Qb6ecNuOub9aPoU06vZemw