Tweak clone to notice if no SCCS/d.files are transferred in the sfio if not it can totally skip the step to trim pending deltas from sfiles.
We previous added dfiles and BitKeeper/etc/SCCS/x.dfile to the SFIO
to make this faster, now skip it entirely if we can.

This is a result of staring a NFS timing of clone.

Fix check to always fix dfile markers in tree.

We were not sending the id_cache in remapped repositories.  This caused
the idcache to get regenerated.

bk: 4b8435bf6wBgAqE4idwfV_Wit40ASQ