A 'cd /tmp/bar/; bk clone /tmp/foo/' would break because of that trailing slash.
A change from bk-4.4 about how clone without a second arguement behaves
caused this bug.  I was previously fixed in the 5.0 codebase because
I must not have noticed the fix was added to -bugfix.
---
This is fixed in 4.6 by the cset -x done in the parent of this cset
which rolls back a bam fix for recomputing urls to bam servers.
When this work is ported forward to the 5.0 line, cset -x the cset -x
to bring it back in. Thanks.

bk: 4baba100_6tiDRjUUxsBtIb-PwnqEw