Perf: fix up clone to be faster on NFS.
This is based on code that Wayne did in 5.x which was based on
doing the same thing by hand by lm.  There is still a little
perf left on the table, if you do this stuff by hand you get
as much as 20% faster.  Wasn't enough to go after it.

We auto-detect NFS and turn on parallelism.  This could be a little
lighter weight if we parsed the mtab but I think that matters only
for us doing regressions.

bk: 4babc564zy-KGfgqREkvW5u5H5Mbkg