Help csetprune by improving perf sccs_adjustSet() and cset_resum().  Each had O(N^2) problem with running serialmap on a O(N) subset of the graph.
The fix has two parts: use graph_symdiff, which computes
the symmetric difference between 2 nodes and limits
how much of the graph is walked; and walk the graph
in a special order which keeps the symdiff walks to
a minimum.

Kudos to Rob for joining in the journey, and reviewing
the work along the way.  Especially looking at
alternative mappings and explore properties of symmetric
difference in relation to corresponding version (the
set of nodes corresponding to a version -- as embodied
in serialmap).  Because of him, the result is math based
with more provability and less reliance on intuition.

Kudos to Wayne for cleaning up the graph_kidwalk and
reviewing it all with an eye toward perf and maintainability.
I like the idea of having symdiff use enum {LEFT, RIGHT}
and have code like "active[LEFT] = 1;" instead of "activeLeft = 1;'.
Also for whipping up the u8 based serialmap which will get
integrated eventually and pave the way for serialmap and friends
(especially sccs_impliedList) making use of graph_symdiff().

bk: 4bd1e476JayN4NusJ_vSQYVU6g3UWA