Portals are fully populated before becoming portals. Attach only works in portals (and checks they are still fully populated). You can't unpopulate comps from a portal.
The idea behind this is that, since we don't have a way to move components,
we'll limit the damage by only allowing attachs to happen in a portal and
requiring portals to be fully populated.

In the future we might loosen this restriction and just check that all the 
components leading to the attach path are populated.

Other changes in this cset:
- We now remove portals with 'bk portal -r'
- 'bk portal' is now not restricted by the license
- doit.sh no longer runs emacs backup files

bk: 4bfc2e79p6MmfZEquGhFpGk-qbRP8A