Big overhaul of progress bars.  The goals are to make the default output look good in the most common use cases, to ensure that -q output really is quiet, and to have the -v progress-bar output not stomp on anything. It is not a goal of this cset to make -v output pretty or to chase down all possible error cases or uncommon use cases; those will get fixed as we hit them in the future.
This work introduces an enhanced progress-bar API (see progress.c) which
includes the ability to delay stderr output while a bk subsystem is doing
its thing and updating a progress bar.  This is a dangerous thing in
that it can make future debugging harder if one isn't aware of it;
see progress_delaystderr() et al.

Thanks go to Wayne and Larry for various ideas and improvements.

bk: 4c0d428acqi5aWoPLzmj7GeaJcVEhQ