Move BitKeeper/deleted into multiple subdirectories.
This avoids large numbers of files in the same directory which
is a big performance problem on some machines, like the <redacted>
NFS filers.

Also fix the 'bk rm' code to ignore files that are already
deleted in the "official" location.  That allows us to use
'bk -rBitKeeper/deleted rm -f' to fanout an existing repository.

bk: 4c10e95dgbmU1_4sDvJBhQRe7ZCXiw