Make clone use lclone by default whenever it can.
Add a new --no-hardlink clone option to prevent this
and a BK_NO_HARDLINK_CLONE override in the environment.

On Windows we don't enable lclone by defualt because the
hardlink semantics on Windows are funky and will cause
problems.  (Unlink fails if any of the hardlinked files
are open.)  The -l option can be use on Windows to requested
(but not require) a hardlinked clone.
Allows this default to be reverted with no-hardlink-clones
and allow -l to require hardlinks.

bk: 4c20c3fbyBt3yddXK_g4fghZrenZgQ