Allow a nested undo to change aliases in the HERE file and add or remove components as needed.
--- (comments from Rick)
REVIEW: the setting ->alias is kind of tricky, and leaving it set is
a little unnerving since it isn't true.  But it seems only ->new
is needed (side note: ->include && ->new is a brain twister since
can ->new be set without ->include?  Whatever, it's not wrong,
just had me go dig through a bunch of code to understand interrelations
(which I had a hand in the nested_init code, so not that hard to refresh)).
Looks good.

---
Fix unrelated AIX regression failure.

bk: 4c4b6c9fIr8QsXtbvT_8-TUJEJPcig