Fix a problem where nested_populate() could end up with a url from the urllist that points at myself and end up thinking it is OK to unpopulate.
Fixed a similar bug in BAM where the BK_BAM_SERVER_ID may have
been incorrect from a component on a bkd.
---
The return code '33' is used to connect the dots by making use
of an otherwise illegal return code bit combination.  The
remote_bk() function is used to connect the local world to
the remote world and has a bitmap of return codes described
in the comment before the function.  0x21 (ie, 33) can't otherwise
happen, so was used for this special purpose of communicating
the same repo id was found.

bk: 4cb195ae8uAEckON9gCdEUEYLUk08w