Rewrite sccslog to use less memory.
It also becomes compatible with the code to put delta
structs in a single memory array.
---
The ouput of sccslog -D is really problematic.  Look at
  bk sccslog -D src/zone.c   (the output is useless)

The output is:
filename list1:
   comment list1
filename list2:
   comment list2

The order of the filename lists depends on a hash.
The order of filenames within a list depends on a hash.
The order of comments under a group of files
depends on a hash.

This code didn't change anything here, just reproduces the old
broken behavior.

bk: 4cbcb201wdr4jR1orDYRrivU6xJ7fw