Fix a bug seen at <redacted>.
A partial_check failed because we stripped a pending rename
without moving the file back to the original localtion.
This is expected and check returns 2 to indicate it autofixed
the problem and that we should retry.

However the retry fails because the file list for partial_check
is no longer valid.  Force a full check in this case.  It is 
safer anyway.

bk: 4ccf3017thqI5fAMQkMibhpRM3vy9Q