Fix "in grace period" warning to not give stale warnings
A <redacted> they have been seeing warnings about a license being
expired and in the grace period for weeks have it has been
updated here.  This is mostly caused because they have lots
of machines and the leases are saved per-machine.  We a
use moves to a machine they haven't touched since before
we updated our leaseDB they have a valid license so the
commit is allowed, but they get a warning because the
license has expired.

Now when we are about to print that warning we refresh the
lease to make sure it is current and check again.

bk: 4cd998b0JteaXyiazP0AliKsaWWSmg