NESTED: fix up r2c and guis in comps.
This is a questionable cset that sprinkles -s. in a few more
places, 
makes revtool on a file turn on -s. implicitly,
makes fm3tool more or less run -s. as it is going from a file
to csets (and can't see the product csets, they are hiding in
RESYNC).

Without this cset, merging in components is pretty darn busted.

bk: 4cd9ec66q7O3gGxJ2309VfNWdIp6iA