Add some trivial indentation to the cmd log.
cmdlog command syntax updates:
  Add an optional search pattern argument in cmdlog.
  Make new output format show just the commands.
  Make -v do the legacy format.
  Add --user, --date, --version, --exit to selectively add back the
  things that are filtered in the new format.
  Update manpage.

Add wayne's suggestion to quote things before logging:

 Currently, if you do

  bk echo 'foo bar'

 the log will have

  echo foo bar = ...

 with this, the you'll get

  echo "foo bar" = ...

Fix write_log so that it only does one fprintf per entry instead of two
and make the FILE* be unbuffered.  The hope here is that we close the
window on the possible race Wayne observed with clone -j8 losing the
odd sfio log message.  This is still imperfect when the log file is
an NFS file _and_ we fire off a whole bunch of commands at once (each
one tries to log in cmdlog_start() and messages can get lost).  We're
okay with that (for now).

Add a release-note blurb about cmdlog changes.

bk: 4d262707kryna2TOpFD8iYIyH2jOSA