Rewrite rset to always run in md5key (-5) mode. This is a major performance win because previously rset needed to init every sfile being printed to get the rev numbers.
In the process Rick and I added a number of additional
tricks and fixes.  Renamed components are handled better
for example.

One hard part is that since we are no longer loading sfiles 
we can't identify and skip sfiles that are gone'd.   Instead
we use GET_SKIPGONE to compare the csetfile weave to the gone
file and for those keys that match we stat() the sfile to see
if it is really gone.  In that case the files are skipped.

This gone handling can be disabled with --show-gone.

bk: 4d267ab3c_s5NtJktYBe3fsQVVdbhg