Allow users to quit out of resolve and restart later on without aborting the whole thing.
This involves many changes to how pulls are done. Normal
pulls will now automatically resolve as many components
as possible and leave all conflicts until the end. The idea
is to preserve nested's "one-pass" behaviour while at the same
time giving the user the opportunity to quit out of it. 

Since now we do all the data transfer _first_, we can immediately
disconnect before dropping to the interactive part. This lets us
unlock the bkd side as soon as we're done transferring data. Before
this cset, a user leaving a resolver conflict prompt unattended could
keep the remote repository locked indefinitively.

More work is needed in resolve.c:resolve_cleanup() because removing the
RESYNC directory of a component just should not be done. A user typing
'a' at the resolver prompt, or a trigger exiting 1 can still leave a 
screwed up repository, but it will not propagate and it can be fixed
by a "bk abort".

bk: 4d276701NoiOmuRInWUZLznEx_V2zg