Add an extra parameter to getServerInfo(). The idea is that if you pass a hash to it, it will fill it in with the remote bkd's environment rather than change the local environment.
This will be useful when we are talking to multiple BKDs, for
instance in the pull-gate RTI, or pull-urllist, etc. Basically
whenever we switch BKD's mid-operation, we shouldn't change the
client's environment.

All the calls to getServerInfo() have been changed to pass a NULL
for the hash, so the code remains functionally equvalent. The 
pull-gate RTI will be the first user of the new hash argument and
bugs found while coding that will be fixed in this RTI.

bk: 4d34cd216UO92xNwpQrgIujk2UtE4A