Add an extended version of Rick's t.nested-triggers-dev testcase.
For a series of commands we enable all triggers and report what
triggers get run from which directories and also check the
which repository locks are present.

This required a bunch of fixes in bk to make the various commands
behave identically.

we were inconsistant about trigger output in places.  
We originally suppressed passing trigger output when -q was passed.  
When we added -v in addition to -q, we sometimes translated !-q  to -v. 
So in some cases we only show trigger output with -v and others 
only without -q.

I also had to fix 'bk lock -l' to get correct output and
fix locking in rclone.

This still shows that we fail to downgrade nested locks before calling
post triggers.

bk: 4d4756c9ql0AOprEnC1EtIP8LqXiyw