Change hash_tofile() routines to webencode() any funky characters in keys so that we can store hash's with arbitrary keys.
To make this efficient webencode() is changed to write directly
to a stdio stream.  This is often a real filehandle, but fmem
is used in cases where the output needs to be used.

Fmem is also used in hash_fromStream() to accumlate data.
This is faster in tests.

bk: 4d495648dfbmZOpm9L44Z_tIDBiX2w