Allow a nested abort to run undo in a gate.
Since we don't allow undo or unpopulate in a gate a
nested abort doesn't really work.  This means a failing push
to a gate will be borked.

I tried to fix this by hooking on the _BK_TRANSACTION
flag.
---
RIDER: do not allow bk here set -f in a portal
or gate anymore.

bk: 4d7e86604iNaJkqDJYjHkY-gbR4xaA