Simplify the converge code for the attribute file.
When two files are created in parallel this code picks the newest
and deletes it and then copies that newer file to 1.0.1.1 so
it can be merged with the older file.

For attr we don't need a 1.0.1.1 because we don't really merge
the data at all, we just write the new values.  So we skip
this part.  This makes the graph much simpler and avoids
bugs in bk-4.x when trying to merge these funny deltas.

bk: 4d9c8d8675MOaFORfUBc9ofQBoXZOQ