So aborting a resolve looks scary:
.                                  100% |==============================| FAILED
gcc/ChangeSet: marked delta 1.4 should be in ChangeSet but is not.
	1.4 -> bk/ob@bk_regression.bk/dirac.bitmover.com|gcc/ChangeSet|20110316222231|39542
gcc/ChangeSet: marked delta 1.3 should be in ChangeSet but is not.
	1.3 -> bk/ob@bk_regression.bk/dirac.bitmover.com|gcc/ChangeSet|20110316222229|26544
newcomp/ChangeSet: marked delta 1.1 should be in ChangeSet but is not.
	1.1 -> bk/ob@bk_regression.bk/dirac.bitmover.com|newcomp/ChangeSet|20110316222227|13418
check: File newcomp/ChangeSet doesn't have any data that matches the local ChangeSet file.
This file was likely copied from another repository

This cset adds a friendly message to the end:

-----------------------------------------------------------------
It looks like there was a pull in progress. These errors could
be caused by the components being out-of-sync with the product.
Try running 'bk resolve' to finish the pull, or a 'bk abort' to
cancel it.
-----------------------------------------------------------------

Hopefully it'll lower support.
--- rick says ---
Larry noted that abort in a component is less than graceful.
This is a hack fix until someone who understands the resolve
picture puts parts in place.
This takes the return of a component resolve and sees if
the product RESYNC is still there.  If not, it exits
resolve using the longjmp hook set up for such hacks as this.

bk: 4dc3226ejrRIIs9SLpS1GGSpXai07Q