Fix pull failure reported by ntp.org.  They had a new sfile in editd and modified in the same path as a change to an older sfile that had been renamed in the local repo.
In nameOK(), check that the keys match before complaining and
setting the opts->edited flag.

Also, in t.a.progress-bars remove tests for this situation
that were assuming this was a real conflict/error

An equivalent test to show this now works has been added to
t.resolve.

bk: 4dd3d97duMO8_Jj-xRQoa2mixrF3yw