Before bringing over any old ChangeSet file, run findkey on the parent(s) the -@ named repos to make sure that the desired tip key is present.
The hope is to prevent the noisy mess that is a repository with sfiles that
are ahead of the ChangeSet file.
---
Add a --use-older-changeset option.

This is for those unlikely situations where a customer has lost
a ChangeSet file and the only place to repair from is behind
(does not contain the needed tip key necessary to effect the
repair).

This option remains undocumented because it has the potential
to make a broken repository more broken (at least the check
messages are scarier).

bk: 4e98e975g8tMSQw22sbV6-ojWhPZvw