If we are doing an rclone then use sfiles_clone -m unless the other side doesn't hable sfio -m or if we are doing an rclone -rREV.
The rclone -r part is important, without disabling this for -r we
fail 3-4 regressions.  I did not look to see if we were smart in
clone -r.
---
Get -m to work if -r, though in a sense clone should remove
-m if -r to match this.  Or really, clone should unlink HERE
if -m.  That's a different problem related to clone -r and
aliases (similar to RTI about pull-r-aliases)

bk: 4eab09e1zq38ZkSKoAvJnM98QeBFGA