In bk-5.3, partition was enhanced to handle duplicate file basenames with same RANDOM bits.  To be backward compat, a version was added to the stored data in BitKeeper/log/partition.
Any existing partitions would not have this version.  When
the stored data was written, it would core dump because
it assumed all variable would have had values read in.
Fixed it to allow variables to not have values.  And to
not ever write out the version if version is 1 (to be compat
with how the old (bk-5.0) did it).

bk: 4eb01f6c8XHWElOXuPOPVnB4G5mbzA