PERF: make clones go faster
The idea was to make the checkouts happen as close to the unpacking
of the sfiles as possible and to have multiple checkouts running.
Originally for NFS but it turns out it helps on local disk as well.

Rick added support to ignore symlinks on windows.

bk: 4ebb1801ef6qDJcpbPRWWlvV-yR1aw