Always have the graph be compressed (no unused serials). This is maintained by having stripdel do scompression as part of removing D_GONE nodes.  And having csetprune use stripdel.
This means that loops which walk the table do not need
to check for holes.  As a safety net, holes are checked
for inbound (mkgraph for ascii sfiles) and outbound
(delta_table for both ascii and binary).

Since ascii sfiles can have holes, they will be scompressed
on the fly.  Binary files can't have holes.

As part of all of this, drop 'bk _scompress' and block
'bk rmdel' being used on a BK sfile.

bk: 4ec3c6cbBjk8BaChWPw6qa5uIZLwxw