Simplify the locking in unique.c This was causing bad bugs for the Sun guys when the same file is being written over NFS, but we have the lockfile in /tmp. Put keys in a per host location.
Also keep the .bk_keys file open for the duration of any process
that creates a new key (admin, delta, commit, ...)  This saves
rereading that file over and over again.  This require the file
be explicitly closed in a couple places to avoid deadlocks.

This cset generally a net perf win, even on NFS.

Also remove the lock_dir() function.  It is almost exactly
the same as TMP_PATH which calls nt_tmpdir() and we don't
need two functions like this.
---
Locking: make unique be noisy.
This fixes a long standing bug in sccs_lockfile that made it never
be noisy. (lm)

bk: 4ede91a127m0fZGJIjnG4TVL0CjMmQ