Change binary sfile format to use the new fgzip.c compressed file format that verifies checksums when reading.
The headers in fgzip.c and fchksum.c explain the new file format
wrappers.  The comment on slib.c:bin_mkgraph() explain the file
format stored in these layers.

This code enables a the features we are adding the the future
where sfiles can be paged in from disk and commit can update
the ChangeSet file in-place without reading the entire file.

bk: 4ee63fb9vMddiewv2AhJptRUX9U9lA