Add to check a check for duplicate -i or -x in the graph. Wanted so we can check existing customers' repos to know if handling duplicates is something we need to address.
The underlying issue is the transform for going between
SCCS bookkeeping (-i are used for merge) and BK bookkeeping (the
merge edge acts like the parent edge).  The core transform
collapses out duplicates, so a SCCS -> BK -> SCCS could result
in something different.  Choices are to support listing duplicates
in the BK domain (and therefore back) giving identical results,
or refusing to do the transform if any collapses; instead having
the customer transform all their repos (a new root operation),
then do the transform to BK.  This is similar to how partition
doesn't work with shortkeys, expect that partition has no map
back to the original domain.

Wayne advocated for the extra-transform/newroot and suggested
this check as a way to size up it might mean for customers.

bk: 4f172de4l-NBLdxZRcyXhSquNxWDsQ