BAM: bugfix for wagerworks/IGT
The problem is that if a file has been goned it is not in the
idcache.  So if you delete the s.file, gone it, create another
one there, you now have two rootkeys that point at the same
location.  Without one in the idcache sfiles_bam will list them
twice which causes bam check to barf.

The fix is to pass in the goneDB to key2path and check if the key
is goned.  If it is, then we need to be more careful and check
that the file is the right file, not a newer one.  So we init any
file marked as goned.

bk: 4f2ad5c1AlqoWw824MTLFDhc-WXb6w