Remove D_CHKSUM & D_SORTSUM
This is an order1 file format change and also making maintaining the
in-memory state because we no longer have extra flags that have to
be set correctly or they change the meaning of the fields.

This includes some fixes from Rick, to have 'bk checksum' verify
tag deltas have a zero checksum.

bk: 4f609f9fGIWemb7DTbfOgHGfdGIKbA