Remove all those DF_* defines. This uses df_opt instead but allows passing dop->flags for SILENT.
Remove mdiff. If we need it I'll make it an
output format for ndiff rather than all the
diff parsing that was here.

Change API of diff_files() so the output is a 'char *',
rather than a FILE * (with "-" meaning stdout). This makes
the diffs smaller when plugging diff_files() into slib.c.

Remove "bk diffs -I" support as it was only used
for mdiff. This also simplifies the code in slib.c.

Honor _BK_USE_EXTERNAL_DIFF env variable that was
lost in deleting libdiff.c

bk: 4f7dd6c1MN3IQoKSZhc3GVBHUAYNaw