Rewrite renumber code to use less message and run in O(n) time
If we remove Revs from the binary file format then we will need to be
able to generate revs on the fly.  This includes writing the ascii
file even if we completely remove revs from bk internal and interface.
This code optimizes renumber so it is done in strictly one pass
with only saving one 32-bit number per delta.

Rick contributed several fixes and tweak and an semi-related
cleanup to libgraph.

bk: 4f7ed5d2HPxMNYkKGdCfMnOfvwPqxg