Use stdio passing INIT files between fuctions in bk.
This is a complement to my cset to use stdio for dF, only for iF the INIT
file data.  In order to make this code change clean, I needed to change
all of takepatch.c to use stdio as well.

The prepares the way for having takepatch stream from stdin.

bk: 4fb4f4926LlJyNOp0QJeMfs29SYgxg