Unique db: /tmp backup for when .bk can't be written.
Add a backup for the uniq keys file in /tmp if we can't write the
one in .bk for some reason.
Union what is in .bk and /tmp (imagine that /home was NFS and it
went away and then came back; or quotas or disk space).
The next time we are able to write the .bk one delete the /tmp one.

Originally by wscott, worked over by lm.

bk: 4fbe5205BBD8TBiZbHRv7GDAad14Pg