Remove support for short bk keys.  A cset from 2004
Remove support for short keys and all code that handled short
keys.  At the moment we believe that only the BitKeeper repository
itself needs this support so this is too much baggage to carry around.

A cset with shortkeys will fail in sccs_init and tell the user
to call bk support.

bk: 5005b4eahmGvl8JXVJWH_bASdMMerg