Add a --elide option to rset to omit files that have no diffs. Make difftool use it.
It follows bk-good-enough guidelines and doesn't do
a total sort. Instead, it sorts component paths and 
files within components (deep nests are not handled
inline, a/b/c might come _after_ a/z if a/b/c is a component
and a/z is a file).

bk: 501ac9b7wg89h-f-zhIOgnQ8T-q0WA