Require a port source to be a gate. The intent is to have the user declare that the source repo is important and will be findable and long lived.
<redacted> has ported data out of a product then lost the repo.
Later, when that data gets ported back into the product, it
is blocked.  And there is no matching product cset to be found.
This should lessen that happening.  It still can because a
person can make a repo a gate, then port from it and delete it.

bk: 502402bfOwpNlfFnQI7FjO_3YU0Vkw