nested: make bk repair work in nested trees
It boiled down to respecting TIP if no changeset file and adding
ROOTKEY to the urls.

Demonstrate that we won't find changesets in a standalone.  It's
a bug related to the fact that we md5key the path to the repo (which
supports my belief that comp/BitKeeper/log/TIP should not have the
paths).

This also taught me if we are repairing a file from a file w/
pending deltas we barf (not a prob for this RTI).

bk: 5050cc8cafoULXiHDZHag18i82f7Kw