Add ranges support to bk r2c. Now you can do
  bk r2c -r<low>..<high> <file>

and get back a list of ChangeSet revisions separated
by commas. E.g. "1.5,1.4,1.3". Poly is handled using
the poly db.

The file's cset marks are gotten over a single pass on the graph.
The cset revs are gotten in a single pass on the weave.
For poly we work harder and just run poly_r2c() for each
file delta in the range. It could be optimized if it ever becomes
a problem.

bk: 50770a40bdFfffPscrNRZHdNzM4UTQ