<redacted> saw a message from checkAll():   delta $SOME_KEY missing? This happens when a annotate -hR of the main cset file is removed from a hash of the resync cset file.
A problem can occur when the main cset file has poly
(specifically, the same deltakey twice).  The first
subtraction works and the second trips the error.

One solution is to use checkAll() from stable, which does
the diff by doing a annotate -hR<repo-tipkey>..  which does
the diff without needing to do the diff in checkAll. 
It also puts the results in a hash, so dups fold.

The solution chosen is to just s/sort/sort -u/ to remove the dup.

Also add in errors detection for commit and pull to stop poly
from entering into the repo.  And give support a way to address
it when it does: putting a signed line into BitKeeper/etc/ignore-poly

bk: 50b69a1c2pMEZmeCdpfMPlaFtxFNFQ