Tweak the on-disk poly db data to include the timestamp of the oldest delta in the cset. This is only for perf reasons, which isn't yet an issue because there are no users of poly created by port. Since it adds to the data being stored, we're doing it now.
We know a cset can't contain anything newer than the
tip, but it can include things that are older than
the root(s).  This lets us know a cset is not in the
range by being older than the oldest that is in the range.

bk: 50fda86bOO885ElHV4qQf4LGI6ZjRA