Change how the repository file formats are selected. We now have the rule that a repository format is stable and only changes on clone. Clone gets two new options:   --upgrade-repo    # switch to the lastest and greatest   --compat	    # switch to the oldest supported format
and other than clone the intention is that file formats will not
change nor should you get a repository with files in a mix of
different formats.

This also changes the bSFILEv1 feature flag on disk to BKFILE
which will cause a hickup inside of bitmover.

bk: 50fec664m7T5JxkgLYyH3U3NWp_TtQ