The latest chapter in the features saga
This a major rewrite on the code to manage BitKeper/log/features.
The main point is the never allow the feature in components to
get out of sync with each other.  This is done by only having one
copy of the list in the product.  Also the features parsing is simplified
by having only one piece of code that parses the features file and
features sent over the protocol.
Clone was changed to be more careful when upgrading and downgrading repositories.
Now the correct features are calculated and then these are compared to
the features enabled by the remote of the connection.  This means upgrades
and downgrades happen correctly and component will correctly get set to the right
format even when --upgrade-repo is not used.

bk: 510c369cfBoWgvwSzqet7-i0EAf82g