External heap files and binary weave encoding.
This allows CSET files to write the heap in a separate h.ChangeSet
file from the s.ChangeSet file.  This allows s.ChangeSet to be
1/10th of the original size and be faster to rewrite.  The heap file
doesn't need to be rewritten for pull or commit.  The new data
can just be appended to the end of the heap.

In addition the CSET weave is moved out of the sfile and put into
the delta table as pointers into the heap.  The new encoding is
much smaller than the ascii weave and by linking to the table is can
be accessed in arbitrary order or only the activated serials need
to be parsed.  Also this moves more data into the append-only data
structure.

This cset is just a stepping stone to the final design in the next
cset.  If includes lots of feedback and code from Rick and Oscar.

bk: 512e6495XivSmxNzf8UrSA1f5Iq0iA