Don't call utime() on a sfile in a remapped tree.
BK has lots of code to mess with sfile timestamps for GNU make.
In a remapped tree this isn't needed and just adds more FS operations
over NFS.  So we turn this off in a remapped tree.
---
By not forcing timestamps on s.ChangeSet we exposed a bug
where the TIP cache was not being maintained correctly.
Fixup the caching and add an assertion to catch places where
it doesn't get updated in the future.

bk: 514b46f9qmA6bfxvmTc8qYUrS8Sk6w