BitKeeper Emacs Mode Fixes (vc-bk.el)
This fixes an annoyance I have which is that the C-x v L 
command doesn't run "bk changes", instead it runs "bk log" 
in the root of the repository and that prints the deltas 
not only of the ChangeSet file, but also of all the files
in the root of the repository.

This also fixes a bug where the buffer is not overwriten
on a new C-x v l, but rather the output of bk log is just
dumped wherever the cursor happens to be (usually in the
"Show 2x more" line). 

Also, honor vc-log-show-limit by passing the limit variable
to the -%d option in bk changes and bk log.

bk: 51520b5anfCXGkGJtBUMT7BwUutQhw