Sweep a race under the carpet (will fix it in a different RTI).
This fixes that race in macos-x86, where t.compat-5.3 misinterprets
the @OK@ sent by clone for the @OK@ sent by "bkd nested abort" and
doesn't print the error. Rather than fixing the race (which is dangerous
before a release) we sweep it under the carpet and just make aborting
locks not print an error.

bk: 515afdcciSh8RYqm-9C-cwoE3u6umg