Performance fix for 'bk clone --upgrade-repo' (and --compat)
Now instead doing the sfile conversion with a separate call to
'bk -r admin -Zsame', we allow the consistency check to fix the
files.  We were running a full check already.  As a side effect
running --upgrade-repo on an already upgraded repo won't be slow.

Also in here the special case code for bSFILEv1 is removed we
saves a lot of head scratching everytime I need to make changes
to features.c.

bk: 516325efAYrUqD84pqaNFECAadESlQ