Use accessors for c->present and c->pending
This is to get the performance back from
nested_init(). The idea is that we only stat when we actually
need the information. E.g. most of our loops are:

EACH_STRUCT(n->comps, c, i) {
        unless (c->alias) continue;

        if (c->present) ...
}

My thought was that we can get the alias perf all over the place
by being lazy on the present and pending checks.

Lots of help from Wayne in getting the deltakey to be also lazy 
because of pending.

bk: 5171d058MAa-qoX4MT_CDvRSQ0sZHA