Perf fixes for sfiles
Extent the scancomps file to flag each component with 'e' for edited
and/or 'p' for pending.  Also add a scandirs file to each repository with
the same p/e markings for each directory.  The sfiles is expended to use
the scandirs files to only walk a subset of the directories if it is possible.

Also use scancomps in 'bk -Uc' and friends to only touch a the needed
components.

bk: 5175ce4bHOEfT4GbGxez1EraRHstVQ