Add an an assert (enabled with BK_REGRESSION is set) that ensures that we have a write lock before any update to an s-file begins.
This is based on work teased out of lm's wrlock assertion
tree (while dropping a pile of other locking api changes).

The bulk of this changeset is changes to commands that were
running afoul of this assertion OR fixing the invocations of
bk subprocesses that would otherwise deadlock on locks held
by their parent.

bk: 51a8d6f1_KH_5kv3Sg__yOJG-dU3Cw