If we call "system(cmd)" on windows, and "cmd" prints to stdout or stderr, the system() fails silently since we don't have a console.
Since we're trying to ship 6.0, just make a wrapper for system()
called bk_system() that always calls it with &out and &err variables
and alerts us when output there was unexpected.

lm also made it so that all messages go to a GUI now, even on unix.

bk: 51d74c9fMDd7rm3H3KCnCuVBKgc8QQ