Address <redacted> permission failures on 2.ChangeSet
Somehow <redacted> is getting repos with a readonly SCCS/2.ChangeSet
file.  I found that 'bk repair' can do that and fixed that.
I also fixed the code to automatically fix the permissions
in this case.
---
Fix _BK_DEVELOPER so it can be used
and make the permisions fix only happen for users.

bk: 5203d4a5alGSCWgzHjWREi0-gPlaxg