Failed comp pulls can do a full abort and mangle the repo
A repository with a bad BAM server would get mangled when
the commponent pull did a full abort without the toplevel
pull knowing about it.

We also noticed that pre-resolve triggers can have the same
effect.  And that saving PENDING when pre-resolve returns 2
doesn't make any sense in nested so that was removed.

bk: 5204bff4PWxN_VIGeg3KZjoggUFCmg