AndyC found the a BKFILE->compat clone would fail with paging
Clone a BKFILE repo using bk-5 from a bk-6 bkd.  If any file
is big enough to enable paging then the auto-conversion code
in sfio.c would fail with an assertion.  This was caused
because of excess complexity in the s->mem_out code.

Remove the rest of the s->mem_in code.  Most of it was
unused and the complexity was causing bugs and causing people
to not understand what the code is really doing.

bk: 5256c83fYCSx3ZaeUms3m7SnMSV-Aw